+++ io_soulstruct/soulstruct/blender/animation/__init__.py
@@ -1,6 +1,7 @@
 from .import_operators import *
 from .export_operators import *
 from .misc_operators import *
+from .skeleton_operators import *
 from .properties import *
 from .types import *
 from .gui import *